codecrafters submit [skip ci]

+++ app/lib/commands/blpop.go
@@ -7,7 +7,7 @@ import (
 	"github.com/codecrafters-io/redis-starter-go/app/store"
 )
 
-func HandleBLPOP(conn net.Conn, list_key string, timeout int) {
+func HandleBLPOP(conn net.Conn, list_key string, timeout float64) {
 	elements, err := store.Cache.BLPop(list_key, timeout)
 	if err != nil || elements == nil {
 		conn.Write([]byte("*-1\r\n")) // Redis Null Array
